// emm/web_ui/src/api/auth/types.ts
export type Iso8601 = string;
// ...
export type ServiceName =
  | "auth_service"
  | "secret_service"
  | "server_service"
  | "loging_service"
  | "config_service"
  | "docker_registry"
  | string;
// ...
export interface PATCreateResponse {
  token_id: string;
  /** Shown once. */
  token: string;
  name: string;
  expires_at: Iso8601 | null;
}

export interface PersonalAccessToken {
  token_id: string;
  name: string;
  expires_at: Iso8601 | null;
  allowed_services: ServiceName[];
  created_at: Iso8601;
  last_used_at?: Iso8601 | null;
  revoked_at?: Iso8601 | null;
}